feat: robot funtime

+++ src/main/java/ija/project/ijarobots/common/Position.java
@@ -1,16 +1,16 @@
 package ija.project.ijarobots.common;
 
 public class Position {
-    private final int row;
-    private final int col;
+    private final double row;
+    private final double col;
 
-    public Position(int r, int c){
+    public Position(double r, double c){
         this.row = r;
         this.col = c;
     }
 
-    public int getRow(){ return this.row; }
-    public int getCol(){ return this.col; }
+    public double getRow(){ return this.row; }
+    public double getCol(){ return this.col; }
     public double distance(Position p){
         double xSide = Math.abs(p.row - this.row);
         double ySide = Math.abs(p.col - this.col);
